Fixes na de hand van tests Jesse (21-10-2021)

+++ api/src/Entity/Attribute.php
@@ -19,7 +19,7 @@ use Symfony\Component\Serializer\Annotation\MaxDepth;
 use Symfony\Component\Validator\Constraints as Assert;
 
 /**
- * An possible attribute on an Entity
+ * An possible attribute on an Entity.
  *
  * @category Entity
  *
@@ -70,7 +70,7 @@ class Attribute
      * )
      * @Assert\NotNull
      * @Groups({"read", "write"})
-     * @ORM\Column(type="string", length=255, nullable=true)
+     * @ORM\Column(type="string", length=255)
      */
     private $name;
 
@@ -297,7 +297,6 @@ class Attribute
      */
     private $requiredIf = [];
 
-
     /**
      * @var array conditional requiremends for field
      *
@@ -447,12 +446,12 @@ class Attribute
     private $maxDate;
 
     /**
-     * @var string *Can only be used in combination with type file* The maximum allowed file size in KB
+     * @var string *Can only be used in combination with type file* The maximum allowed file size in bytes
      *
-     * @example 32
+     * @example 32000
      *
      * @Groups({"read", "write"})
-     * @ORM\Column(type="float", nullable=true)
+     * @ORM\Column(type="integer", nullable=true)
      */
     private $maxFileSize;
 
@@ -473,24 +472,21 @@ class Attribute
     private $validations;
 
     /**
-     *
-     * Setting this property to true wil force the property to be saved in the gateway endpoint (default behafure is saving in the EAV)
+     * Setting this property to true wil force the property to be saved in the gateway endpoint (default behafure is saving in the EAV).
      *
      * @ORM\Column(type="boolean", nullable=true)
      */
     private $persistToGateway = false;
 
     /**
-     *
-     * Whether or not this property is searchable
+     * Whether or not this property is searchable.
      *
      * @ORM\Column(type="boolean", nullable=true)
      */
     private $searchable = false;
 
     /**
-     *
-     * Whether or not this property kan be used to create new entities (versus when it can only be used to link exsisting entities)
+     * Whether or not this property kan be used to create new entities (versus when it can only be used to link exsisting entities).
      *
      * @ORM\Column(type="boolean", nullable=true, name="allow_cascade")
      */
@@ -778,7 +774,6 @@ class Attribute
         return $this;
     }
 
-
     public function getForbidenIf(): ?array
     {
         return $this->forbidenIf;
@@ -812,10 +807,8 @@ class Attribute
     {
         $this->type = $type;
 
-
-
         // If the attribute type is changes away from an object we need to drop the object
-        if($type != 'object' and $this->object ) {
+        if ($type != 'object' and $this->object) {
             unset($this->object);
         }
 
@@ -990,12 +983,12 @@ class Attribute
         return $this;
     }
 
-    public function getMaxFileSize(): ?float
+    public function getMaxFileSize(): ?int
     {
         return $this->maxFileSize;
     }
 
-    public function setMaxFileSize(?float $maxFileSize): self
+    public function setMaxFileSize(?int $maxFileSize): self
     {
         $this->maxFileSize = $maxFileSize;
 
@@ -1014,30 +1007,29 @@ class Attribute
         return $this;
     }
 
-
     public function getValidations(): ?array
     {
         //TODO: this list of validations is not complete!
         $validations = [];
-        $validations['maximum']             = $this->getMaximum();
-        $validations['exclusiveMaximum']    = $this->getExclusiveMaximum();
-        $validations['minimum']             = $this->getMinimum();
-        $validations['exclusiveMinimum']    = $this->getExclusiveMinimum();
-        $validations['maxLength']           = $this->getMaxLength();
-        $validations['minLength']           = $this->getMinLength();
-        $validations['maxItems']            = $this->getMaxItems();
-        $validations['minItems']            = $this->getMinItems();
-        $validations['uniqueItems']         = $this->getUniqueItems();
-        $validations['maxProperties']       = $this->getMaxProperties();
-        $validations['minProperties']       = $this->getMinProperties();
-        $validations['required']            = $this->getRequired();
-        $validations['enum']                = $this->getEnum();
-        $validations['allOf']               = $this->getAllOf();
-        $validations['anyOf']               = $this->getAnyOf();
-        $validations['oneOf']               = $this->getOneOf();
-        $validations['defaultValue']        = $this->getDefaultValue();
-        $validations['nullable']            = $this->getNullable();
-        $validations['mustBeUnique']        = $this->getMustBeUnique();
+        $validations['maximum'] = $this->getMaximum();
+        $validations['exclusiveMaximum'] = $this->getExclusiveMaximum();
+        $validations['minimum'] = $this->getMinimum();
+        $validations['exclusiveMinimum'] = $this->getExclusiveMinimum();
+        $validations['maxLength'] = $this->getMaxLength();
+        $validations['minLength'] = $this->getMinLength();
+        $validations['maxItems'] = $this->getMaxItems();
+        $validations['minItems'] = $this->getMinItems();
+        $validations['uniqueItems'] = $this->getUniqueItems();
+        $validations['maxProperties'] = $this->getMaxProperties();
+        $validations['minProperties'] = $this->getMinProperties();
+        $validations['required'] = $this->getRequired();
+        $validations['enum'] = $this->getEnum();
+        $validations['allOf'] = $this->getAllOf();
+        $validations['anyOf'] = $this->getAnyOf();
+        $validations['oneOf'] = $this->getOneOf();
+        $validations['defaultValue'] = $this->getDefaultValue();
+        $validations['nullable'] = $this->getNullable();
+        $validations['mustBeUnique'] = $this->getMustBeUnique();
 
         return $validations;
     }
@@ -1045,25 +1037,63 @@ class Attribute
     public function setValidations(?array $validations): self
     {
         //TODO: this list of validations is not complete!
-        if(array_key_exists('maximum',$validations)){           $this->setMaximum($validations['maximum']);}
-        if(array_key_exists('exclusiveMaximum',$validations)){  $this->setExclusiveMaximum($validations['exclusiveMaximum']);}
-        if(array_key_exists('minimum',$validations)){           $this->setMinimum($validations['minimum']);}
-        if(array_key_exists('exclusiveMinimum',$validations)){  $this->setExclusiveMinimum($validations['exclusiveMinimum']);}
-        if(array_key_exists('maxLength',$validations)){         $this->setMaxLength($validations['maxLength']);}
-        if(array_key_exists('minLength',$validations)){         $this->setMinLength($validations['minLength']);}
-        if(array_key_exists('maxItems',$validations)){          $this->setMaxItems($validations['maxItems']);}
-        if(array_key_exists('minItems',$validations)){          $this->setMinItems($validations['minItems']);}
-        if(array_key_exists('uniqueItems',$validations)){       $this->setUniqueItems($validations['uniqueItems']);}
-        if(array_key_exists('maxProperties',$validations)){     $this->setMaxProperties($validations['maxProperties']);}
-        if(array_key_exists('minProperties',$validations)){     $this->setMinProperties($validations['minProperties']);}
-        if(array_key_exists('required',$validations)){          $this->setRequired($validations['required']);}
-        if(array_key_exists('enum',$validations)){              $this->setEnum($validations['enum']);}
-        if(array_key_exists('allOf',$validations)){             $this->setAllOf($validations['allOf']);}
-        if(array_key_exists('anyOf',$validations)){             $this->setAnyOf($validations['anyOf']);}
-        if(array_key_exists('oneOf',$validations)){             $this->setOneOf($validations['oneOf']);}
-        if(array_key_exists('defaultValue',$validations)){      $this->setDefaultValue($validations['defaultValue']);}
-        if(array_key_exists('nullable',$validations)){          $this->setNullable($validations['nullable']);}
-        if(array_key_exists('mustBeUnique',$validations)){      $this->setMustBeUnique($validations['mustBeUnique']);}
+        if (array_key_exists('maximum', $validations)) {
+            $this->setMaximum($validations['maximum']);
+        }
+        if (array_key_exists('exclusiveMaximum', $validations)) {
+            $this->setExclusiveMaximum($validations['exclusiveMaximum']);
+        }
+        if (array_key_exists('minimum', $validations)) {
+            $this->setMinimum($validations['minimum']);
+        }
+        if (array_key_exists('exclusiveMinimum', $validations)) {
+            $this->setExclusiveMinimum($validations['exclusiveMinimum']);
+        }
+        if (array_key_exists('maxLength', $validations)) {
+            $this->setMaxLength($validations['maxLength']);
+        }
+        if (array_key_exists('minLength', $validations)) {
+            $this->setMinLength($validations['minLength']);
+        }
+        if (array_key_exists('maxItems', $validations)) {
+            $this->setMaxItems($validations['maxItems']);
+        }
+        if (array_key_exists('minItems', $validations)) {
+            $this->setMinItems($validations['minItems']);
+        }
+        if (array_key_exists('uniqueItems', $validations)) {
+            $this->setUniqueItems($validations['uniqueItems']);
+        }
+        if (array_key_exists('maxProperties', $validations)) {
+            $this->setMaxProperties($validations['maxProperties']);
+        }
+        if (array_key_exists('minProperties', $validations)) {
+            $this->setMinProperties($validations['minProperties']);
+        }
+        if (array_key_exists('required', $validations)) {
+            $this->setRequired($validations['required']);
+        }
+        if (array_key_exists('enum', $validations)) {
+            $this->setEnum($validations['enum']);
+        }
+        if (array_key_exists('allOf', $validations)) {
+            $this->setAllOf($validations['allOf']);
+        }
+        if (array_key_exists('anyOf', $validations)) {
+            $this->setAnyOf($validations['anyOf']);
+        }
+        if (array_key_exists('oneOf', $validations)) {
+            $this->setOneOf($validations['oneOf']);
+        }
+        if (array_key_exists('defaultValue', $validations)) {
+            $this->setDefaultValue($validations['defaultValue']);
+        }
+        if (array_key_exists('nullable', $validations)) {
+            $this->setNullable($validations['nullable']);
+        }
+        if (array_key_exists('mustBeUnique', $validations)) {
+            $this->setMustBeUnique($validations['mustBeUnique']);
+        }
 
         return $this;
     }
@@ -1139,7 +1169,4 @@ class Attribute
 
         return $this;
     }
-
-
-
 }
